Changed argument in Entity constructor HIERARCHY_INDEX from short to int

+++ src/engine/audio/SoundSource.java
@@ -15,7 +15,7 @@ public class SoundSource extends Entity
     private float volume;
     private float pitch;
 
-    public SoundSource(String name, short HIERARCHY_INDEX, SoundClip clip, boolean loops)
+    public SoundSource(String name, int HIERARCHY_INDEX, SoundClip clip, boolean loops)
     {
         super(name, EntityType.SoundSource, HIERARCHY_INDEX);
         this.clip = clip;
