+++ src/engine/core/Camera.java
@@ -11,7 +11,7 @@ public class Camera extends Entity
     public float nearPlane;
     public float farPlane;
 
-    public Camera(String name, short HIERARCHY_INDEX)
+    public Camera(String name, int HIERARCHY_INDEX)
     {
         super(name, EntityType.Camera, HIERARCHY_INDEX);
         updateViewport(1.0f, -1.0f, 1.0f);
