+++ src/engine/rendering/TextMesh.java
@@ -35,7 +35,7 @@ public class TextMesh extends Entity
     private final ScreenSpace2dShader screenSpace2dShader;
     private final float BASE_SCALE = 0.01408450704225f;
 
-    public TextMesh(String name, short HIERARCHY_INDEX, String assetKey, boolean isLocationAnchored)
+    public TextMesh(String name, int HIERARCHY_INDEX, String assetKey, boolean isLocationAnchored)
     {
         super(name, EntityType.TextMesh, HIERARCHY_INDEX);
         this.font = (Font) Scene.assets.getAssetFromPool(assetKey);
